requirements.txt
Django==4.0.4
djangorestframework==3.13.1
djoser==2.1.0
drf-yasg==1.20.0
djangorestframework-simplejwt==5.0.0
django-debug-toolbar==3.4.0
factory-boy==3.2.1
pytest==7.1.2
pytest-django==4.5.2
pytest-mock==3.7.0

